redux persist
installing, connecting and replacement of localStorage
