+++ client/src/index.js
@@ -2,7 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App/App.js';
 import { Provider } from 'react-redux'
-import { store } from './store/index'
+import { store, persistor } from './store/index'
+import { PersistGate } from 'redux-persist/integration/react'
+import Loader from './components/App/Loader.js';
 
 // ReactDOM.render(
 //   <React.StrictMode>
@@ -14,7 +16,9 @@ import { store } from './store/index'
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
     <Provider store = {store}>
-        <App />
+        <PersistGate loading={<Loader />} persistor={persistor}>
+            <App />
+        </PersistGate>
     </Provider>
 )
