+++ client/src/store/index.js
@@ -1,8 +1,10 @@
-import {createStore, combineReducers} from 'redux'
+import { createStore, combineReducers } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import { authReducer } from './authReducer'
 import { userReducer } from './userReducer'
 import { usersListReducer } from './usersListReducer'
+import { persistStore, persistReducer } from 'redux-persist'
+import storage from 'redux-persist/lib/storage'
 
 const rootReducer = combineReducers({
     isAuth: authReducer,
@@ -10,4 +12,12 @@ const rootReducer = combineReducers({
     users: usersListReducer
 })
 
-export const store = createStore(rootReducer, composeWithDevTools()) 
+const persistConfig = {
+    key: 'root',
+    storage,
+  }
+
+  const persistedReducer = persistReducer(persistConfig, rootReducer)
+
+export const store = createStore(persistedReducer, composeWithDevTools())
+export const persistor = persistStore(store)
